update permission of groupType, check role admin and go

+++ FDS/Services/DocTypeRepository.cs
@@ -1,4 +1,5 @@
 ﻿using FDS.Data;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using System.Xml.Linq;
 
@@ -8,19 +9,41 @@ namespace FDS.Services
     {
         private readonly FDSDbContext _context;
         private readonly GetUser _getUser;
+        private readonly UserManager<ApplicationUser> _userManager;
 
-        public DocTypeRepository(FDSDbContext context, GetUser getUser) 
+        public DocTypeRepository(FDSDbContext context, GetUser getUser, UserManager<ApplicationUser> userManager) 
         {
             _context = context;
             _getUser = getUser;
+            _userManager = userManager;
         }
+        public async Task<bool> checkAccountPermission()
+        {
+            var user = await _getUser.user();
+            var adminRole = await _userManager.GetRolesAsync(user);
+            if (adminRole.Any(a => a.ToUpper() == "ADMIN"))
+            {
+                return true;
+            }
+            var listAccountPermission = await _context.AccountPermissions.ToListAsync();
+            if (listAccountPermission.Any(a => a.UsreName == user.UserName))
+            {
+                return true;
+            }
+            return false;
+        }
+
         public async Task<DocType> Add(DocType type)
         {
+            if (await checkAccountPermission() == false)
+            {
+                return null;
+            }
             var user = await _getUser.user();
             var AllType  = await _context.DocTypes.ToListAsync();
             if (AllType.Any(doc => doc.DocumentType == type.DocumentType))
             {
-                // Tên của Doc mới trùng với một Doc đã có
+                // Tên của type mới trùng với một type đã có
                 return null;
             }
 
@@ -54,12 +77,20 @@ namespace FDS.Services
 
         public async Task<bool> Delete(int id)
         {
-            var result = await _context.DocTypes.Include(q => q.GroupTypes).SingleOrDefaultAsync(a => a.Id == id);
+            if (await checkAccountPermission() == false)
+            {
+                return false;
+            }
+            var result = await _context.DocTypes
+                                    .Include(q => q.GroupTypes)
+                                    .Include(d => d.Documents)
+                                    .SingleOrDefaultAsync(a => a.Id == id);
             if (result == null)
             {
                 return false;
             };
             _context.RemoveRange(result.GroupTypes);
+            _context.RemoveRange(result.Documents);
             _context.Remove(result);
             await _context.SaveChangesAsync();
             return true;
@@ -67,14 +98,26 @@ namespace FDS.Services
 
         public async Task<List<DocType>> GetAll()
         {
-            var result = await _context.DocTypes.ToListAsync();
+            if (await checkAccountPermission() == false)
+            {
+                return null;
+            }
+            var result = await _context.DocTypes
+                                    .Include(q => q.GroupTypes)
+                                    //.Include(d => d.Documents)
+                                    .ToListAsync();
             return result;
         }
 
         public async Task<DocType> GetById(int id)
         {
+            if (await checkAccountPermission() == false)
+            {
+                return null;
+            }
             var result = await _context.DocTypes
                                     .Include(q => q.GroupTypes)
+                                    .Include(d => d.Documents)
                                     .SingleOrDefaultAsync(a => a.Id == id);
             if(result == null)
             {
@@ -85,16 +128,32 @@ namespace FDS.Services
 
         public async Task<bool> Update(DocType type, int id)
         {
-            var result = await _context.DocTypes.SingleOrDefaultAsync(a => a.Id == id);
+            if (await checkAccountPermission() == false)
+            {
+                return false;
+            }
+            var result = await _context.DocTypes.Include(q => q.GroupTypes).SingleOrDefaultAsync(a => a.Id == id);
             if(result == null) 
             {
                 return false; 
             };
-            result.DocumentType = type.DocumentType;
-            result.Note= type.Note;
+            result.DocumentType = type.DocumentType ?? result.DocumentType;
+            result.Note = type.Note ?? result.Note;
             result.Createtor = result.Createtor;
             result.CreateDate = result.CreateDate;
-            _context.DocTypes.Update(result);
+
+            //cập nhật  GroupType
+            foreach (var updatedGroupType in type.GroupTypes)
+            {
+                var _result = result.GroupTypes.FirstOrDefault(gt => gt.Id == updatedGroupType.Id);
+
+                if (_result != null)
+                {
+                    _result.GroupId = _result.GroupId;
+                    _result.Permission = updatedGroupType.Permission ?? _result.Permission;
+                    _result.DocTypeId = _result.DocTypeId;
+                }
+            }
             await _context.SaveChangesAsync();
             return true;
         }
